// hgmall-frontend/src/api/generated/models/create-product-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TheMall API 文档
 * TheMall 商城系统 API 接口文档
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@themall.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CreateSkuRequest } from './create-sku-request';

export interface CreateProductRequest {
    'shopId': number;
    'name': string;
    'description'?: string;
    'imageUrl'?: string;
    'code': string;
    'skus': Array<CreateSkuRequest>;
    'categoryId': number;
}